Align sprite map components with inverted pattern

+++ src/Cyborg.ContentPipeline/AnimationSet.cs
@@ -7,8 +7,8 @@ namespace Cyborg.ContentPipeline
     public class AnimationSet
     {
         public string SpriteSheet { get; set; }
-        public int FrameWidth { get; set; }
-        public int FrameHeight { get; set; }
-        public IDictionary<string, int[]> Animations { get; set; }
+        public short FrameWidth { get; set; }
+        public short FrameHeight { get; set; }
+        public IDictionary<string, short[]> Animations { get; set; }
     }
 }
